requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-bulma==0.8.1
django-cleanup==5.1.0
django-mathfilters==1.0.0
django-registration-redux==2.8
django-summernote==0.8.11.6
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.6
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.2
gunicorn
